Class "threeJsGeometry"

An object of class "threeJsGeometry" represents `three.js` geometry.

threeJsGeometry(...)

# S4 method for class 'threeJsGeometry'
x$name

# S4 method for class 'threeJsGeometry'
x$name <- value

# S4 method for class 'threeJsGeometry'
show(object)

Arguments

...

Each argument in ... becomes an slot in the new threeJsGeometry.

x

an object of threeJsGeometry

name

slot name of threeJsGeometry

value

value to be assigned

object

an object of threeJsGeometry

Slots

x,y,z

"numeric", specify the x, y, and z coordinates.

rotation

"numeric", specify the rotations in the x, y and z axis in radians.

colors

"character", the colors for each geometry.

type

"charater", the type of the geometry. See availableGeometries.

side

'character', the side for side by side plot in threeJsViewer.

layer

'character', the two layer plot in threeJsViewer.

tag

'character', the tag used to group geometries.

properties

A "list", the properties to control the geometry.

Examples

tjg <- threeJsGeometry()