Plot GRanges xyz data in 2d or 3d

Plot GRanges xyz data with grid or rgl package.

view3dStructure(
  obj,
  k = 3,
  feature.gr,
  atacSig,
  renderer = c("rgl", "threejs", "none"),
  lwd.backbone = 2,
  col.backbone = "gray",
  lwd.maxAtacSig = 8,
  reverseATACSig = TRUE,
  col.backbone_background = "gray70",
  alpha.backbone_background = 0.5,
  lwd.gene = 3,
  coor_mark_interval = 5e+05,
  col.coor = "black",
  show_coor = TRUE,
  coor_tick_unit = 50000,
  label_gene = TRUE,
  col.tension_line = "black",
  lwd.tension_line = 1,
  length.arrow = NULL,
  safe_text_force = 3,
  square = TRUE,
  font = readLines(system.file("extdata", "fonts", "helvetiker_regular.typeface.json",
    package = "geomeTriD")),
  ...
)

Arguments

obj

GRanges object with mcols x, y, and/or z

k

The dimension of plot. 2: 2d, 3: 3d.

feature.gr

The annotation features to be added. An object of GRanges.

atacSig

The ATAC-seq signals. An object of GRanges with scores or an object of track.

renderer

The renderer of the 3D plots. Could be rgl or threejs. The threejs will create a htmlwidgets. If 'none' is set, a list of object will be returned.

lwd.backbone, lwd.gene, lwd.tension_line, lwd.maxAtacSig

Line width for the linker, gene, interaction node circle, the dashed line of interaction edges, the tension line and the maximal reversed ATAC signal.

col.backbone, col.backbone_background, col.tension_line, col.coor

Color for the DNA chain, the compact DNA chain, the node circle, the linker, the tension line and the coordinates marker.

reverseATACSig

Plot the ATAC-seq signals in reverse values.

alpha.backbone_background

Alpha channel for transparency of backbone background.

coor_mark_interval

The coordinates marker interval. Numeric(1). Set to 0 to turn it off. The default value 1e5 means show coordinates every 0.1M bp.

show_coor

Plot ticks in the line to show the DNA compact tension.

coor_tick_unit

The bps for every ticks. Default is 1K.

label_gene

Show gene symbol or not.

length.arrow

Length of the edges of the arrow head (in inches).

safe_text_force

The loops to avoid the text overlapping.

square

A logical value that controls whether control points for the curve are created city-block fashion or obliquely. See grid.curve.

font

A font json file for three.js text geometry. See https://threejs.org/docs/index.html?q=font#examples/en/loaders/FontLoader.

...

Not used.

Value

Coordinates for 2d or a list of threeJsGeometry objects or a htmlwidget.

Examples

obj <- readRDS(system.file('extdata', '4DNFI1UEG1HD.chr21.FLAMINGO.res.rds',
    package='geomeTriD'))
feature.gr <- readRDS(system.file('extdata', '4DNFI1UEG1HD.feature.gr.rds',
    package='geomeTriD'))
tjg <- view3dStructure(obj, k=3, feature.gr=feature.gr, renderer='none',
   length.arrow=grid::unit(0.000006, 'native'))